% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-analysis.R
\name{base_erlang}
\alias{base_erlang}
\title{Base Summary of Erlang Stages for Aquatic Life Stages}
\usage{
base_erlang(fileVec, outList, genos, nGenos, nErlang, times, nTimes, nNodes)
}
\arguments{
\item{fileVec}{Vector of files for analysis}

\item{outList}{List of files, organized by repetition, to write output}

\item{genos}{Genotypes to summarize by}

\item{nGenos}{Number of genotypes}

\item{nErlang}{Number of Erlang stages}

\item{times}{Vector of sampling times}

\item{nTimes}{Number of sampled times}

\item{nNodes}{Number of nodes in the network}
}
\value{
None
}
\description{
This function takes the given aquatic stage and summarizes them by Erlang-distributed
dwell times, writing output to provided folders.
}
\details{
This function is a base function used in \code{\link{split_aggregate_CSV}}.
}
