% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-epiImperial.R
\name{base_female_Imperial}
\alias{base_female_Imperial}
\title{Solve Female Distribution with Imperial Model}
\usage{
base_female_Imperial(params)
}
\arguments{
\item{params}{a named list of parameters (see \code{\link{equilibrium_SEI_Imperial_mosy}})}
}
\description{
This calculates equilibrium females, I, based on humans and epi info from
Imperial model only used in equilibrium_SEI_Imperial_mosy
}
