% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-analysis.R
\name{base_gen}
\alias{base_gen}
\title{Base Summary for Eggs, Larvae, Pupae, Susceptible Females, and Infectious Females}
\usage{
base_gen(fileVec, outList, genos, nGenos, nIDX1, times, nTimes, nNodes)
}
\arguments{
\item{fileVec}{Vector of files for analysis}

\item{outList}{List of files, organized by repetition, to write output}

\item{genos}{Genotypes to summarize by}

\item{nGenos}{Number of genotypes}

\item{nIDX1}{First index to expand over, nE/nL/nP for aquatic stages, 1 for the rest}

\item{times}{Vector of sampling times}

\item{nTimes}{Number of sampled times}

\item{nNodes}{Number of nodes in the network}
}
\value{
None
}
\description{
This function takes a given stage and summarizes them by genotype, writing
output to provided folders.
}
\details{
This function is a base function used in \code{\link{split_aggregate_CSV}}.
}
