% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-analysis.R
\name{base_MQ}
\alias{base_MQ}
\title{Base Summary Function}
\usage{
base_MQ(
  fList,
  oDir,
  sName,
  nodeNames,
  nNodes,
  genos,
  nGenos,
  times,
  nTimes,
  num_repss,
  mean,
  quantiles,
  oDepth
)
}
\arguments{
\item{fList}{File list, all files for this stage, organized by repetition}

\item{oDir}{Output directory}

\item{sName}{Stage signifier}

\item{nodeNames}{Properly formatted vector of node names for printing}

\item{nNodes}{Number of nodes in the simulation}

\item{genos}{Vector of genotypes for the header}

\item{nGenos}{Number of genotypes}

\item{times}{Vector of sampling times}

\item{nTimes}{Number of sampled times}

\item{num_repss}{Number of repetitions from the simulation}

\item{mean}{Boolean, calculate mean or not}

\item{quantiles}{Vector of quantiles to calculate, or NULL}

\item{oDepth}{Max(1, number of quantiles)}
}
\value{
None
}
\description{
This function does the actual calculations for \code{\link{summarize_stats_CSV}}.
It calculates mean and quantiles, writing output to the appropriate folder.
}
