% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-analysis.R
\name{base_MUH}
\alias{base_MUH}
\title{Base Summary for Males, Unmated Females, and Humans}
\usage{
base_MUH(fileVec, outList, genos, nGenos, nTimes, nNodes)
}
\arguments{
\item{fileVec}{Vector of files for analysis}

\item{outList}{List of files, organized by repetition, to write output}

\item{genos}{Genotypes to summarize by}

\item{nGenos}{Number of genotypes}

\item{nTimes}{Number of sampled times}

\item{nNodes}{Number of nodes in the network}
}
\value{
None
}
\description{
This function takes a given stage (males, unmated females, or humans) and
summarizes them by genotype (infection status for humans), writing output to
provided folders.
}
\details{
This function is a base function used in \code{\link{split_aggregate_CSV}}.
}
