% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-analysis.R
\name{base_sum_F}
\alias{base_sum_F}
\title{Base Summary of Infection (SEI) Stages for Adult Females}
\usage{
base_sum_F(fileList, outList, genos, nGenos, nErlang, times, nTimes, nNodes)
}
\arguments{
\item{fileList}{Length 3 list holding 'FS','FE', and 'FI' files for analysis}

\item{outList}{List of files, organized by repetition, to write output}

\item{genos}{Genotypes to summarize by}

\item{nGenos}{Number of genotypes}

\item{nErlang}{Number of Erlang stages}

\item{times}{Vector of sampling times}

\item{nTimes}{Number of sampled times}

\item{nNodes}{Number of nodes in the network}
}
\value{
None
}
\description{
This function takes ALL of the adult female stages and summarized them by
Erlang-distributed latent infection, writing output to provided folders.
}
\details{
This function is a base function used in \code{\link{split_aggregate_CSV}}.
}
