% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-plot.R
\name{base_summarize_humans}
\alias{base_summarize_humans}
\title{Base Function for Human Summary}
\usage{
base_summarize_humans(out, infState)
}
\arguments{
\item{out}{the output of \code{\link[MGDrivE2]{sim_trajectory_R}}}

\item{infState}{type of humans to summarize: 'S','E','I','R'}
}
\value{
a 4 to 6 column dataframe for plotting with ggplot2
}
\description{
This function takes a given infection ('S','E','I','R') status and returns a
summary trajectory
}
\details{
This function is the base function for \code{\link{summarize_humans_epiSIS}},
\code{\link{summarize_humans_epiSEIR}}.

The return object depends on the data provided. If the simulation was only 1 node,
then no \code{node} designation is returned. If only one repetition was performed,
no \code{rep} designation is returned. Columns always returned include: \code{time},
\code{inf}, \code{genotype}, and \code{value}.
}
