% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-batch.R
\name{batch_migration_stage}
\alias{batch_migration_stage}
\title{Internal function to sample and set up data structure for batch migration}
\usage{
batch_migration_stage(SPN_P, rates, move, prob, stage, tmax)
}
\arguments{
\item{SPN_P}{a set of Petri net places}

\item{rates}{a vector of rates for each node}

\item{move}{a movement matrix (where do the batches go?)}

\item{prob}{the probability vector for each individual moving in each batch}

\item{stage}{the life stage (one of 'ELP', 'F', 'M')}

\item{tmax}{maximum simulation time}
}
\description{
Internal function to sample and set up data structure for batch migration
}
