% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-epi-decoupled.R
\name{convert_prevalence_to_eir}
\alias{convert_prevalence_to_eir}
\title{Generally, pathogen prevalence is a more accesible metric for users, but
the Imperial equilibrium function requires an annual EIR.
This function converts a given pathogen prevalence to an EIR}
\usage{
convert_prevalence_to_eir(prevalence, age_vector, ft, params)
}
\arguments{
\item{prevalence}{desired prevalence value}

\item{age_vector}{age distribution of the population}

\item{ft}{proportion treated}

\item{params}{entomological and epidemiological parameters}
}
\value{
a vector of the equilibrium number of humans in each SIS stage
}
\description{
Generally, pathogen prevalence is a more accesible metric for users, but
the Imperial equilibrium function requires an annual EIR.
This function converts a given pathogen prevalence to an EIR
}
