% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-SEMcis.R
\name{cubeSEMcis}
\alias{cubeSEMcis}
\title{Inheritance Cube: ReMEDE (REpeat Mediated Excision of a Drive Element)}
\usage{
cubeSEMcis(
  pF = 1,
  qF = 1,
  rF = 0,
  aF = 1,
  bF = 1,
  cF = 1,
  pM = pF,
  qM = qF,
  rM = rF,
  aM = aF,
  bM = bF,
  cM = cF,
  mmrF = 0,
  mmrM = mmrF,
  pDep = 0,
  qDep = 0,
  rDep = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)
}
\arguments{
\item{pF}{Rate of cleavage during GD process in females}

\item{qF}{Rate of HDR during GD process in females}

\item{rF}{Rate of in-frame resistance generation during GD process in females}

\item{aF}{Rate of cleavage during SEM process in females}

\item{bF}{Rate of SSA during SEM process in females}

\item{cF}{Rate of "V" allele formation from SSA during SEM process in females}

\item{pM}{Rate of cleavage during GD process in males}

\item{qM}{Rate of HDR during GD process in males}

\item{rM}{Rate of in-frame resistance generation during GD process in males}

\item{aM}{Rate of cleavage during SEM process in males}

\item{bM}{Rate of SSA during SEM process in males}

\item{cM}{Rate of "V" allele formation from SSA during SEM process in males}

\item{mmrF}{Rate of MMR in females, driving allelic conversion of "V" into "W"}

\item{mmrM}{Rate of MMR in males, driving allelic conversion of "V" into "W"}

\item{pDep}{Rate of cleavage during maternal deposition}

\item{qDep}{Rate of HDR during maternal deposition}

\item{rDep}{Rate of in-frame resistance generation during maternal deposition}

\item{eta}{Genotype-specific mating fitness}

\item{phi}{Genotype-specific sex ratio at emergence}

\item{omega}{Genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{Genotype-specific female pupatory success}

\item{xiM}{Genotype-specific male pupatory success}

\item{s}{Genotype-specific fractional reduction(increase) in fertility}
}
\value{
Named list containing the inheritance cube, transition matrix, genotypes,
wild-type allele, and all genotype-specific parameters.
}
\description{
The ReMEDE system, put forth by \href{GETLINKWHENEXISTS!!!}{Chennuri and Myles},
is a realized application of \href{https://doi.org/10.1098/rstb.2019.0804}{biodegradable gene drives}
with application of \href{https://doi.org/10.1016/j.celrep.2020.107841}{Small-Molecule Control}
for the removal of drive elements through induceable SSA.
}
\details{
This construct has a
single, autosomal target site that carries gRNA, Cas9, an induceable endonuclease,
and is flanked by direct repeats. There are 7 possible alleles:
\itemize{
 \item W: Wild-type
 \item G: Active GD (gene drive), with inactive but functional SEM (self-elimination mechanism) element
 \item U: Low-cost resistant allele, non-targetable by GD or SEM (R1 in other literature)
 \item R: High-cost resistant allele, non-targetable by GD or SEM (R2 in other literature)
 \item V: "wild-type", product of SEM, non-targetable by GD or SEM
 \item H: Active GD with active SEM
 \item S: Active GD with non-functional SEM
}

"V" alleles are simply a minor allele with the same protein sequence as the
major allele, "W". There is the possibility for allelic conversion of the "V"
allele into the "W" allele by mechanisms such as MMR.

This drive has male and female specific GD and SEM parameters, as well as \strong{correct}
maternal deposition. This is the first instance of an inheritance cube built
in this manner with alleles properly specified. There are no dosage effects
modeled (i.e., having two GD alleles increasing or decreasing the GD rates).

Gene-drive and SEM parameters (p*, q*, r*, a*, b*, c*, mmr*) are
all rates and values must fall in the range of [0, 1], inclusive.

Population parameters (phi, xiF, xiM) are either NULL, for default values, or
named vectors, where the names must match the names in the inheritance pattern.
These parameters are also rates and fall in the range of [0, 1], inclusive.

Omega and s may also be NULL or named vectors, but are weights, and thus fall
in the range [0, inf).

Eta, the male mating weights, may be NULL for default values, or applied as a
list of vectors. If the vectors are length 2 (e.g., c(maleGeno, matingWeight)),
then that weight is applied to all female genotypes for that specific male genotype.
Alternatively, vectors of length 3 may be supplied (e.g., c(femaleGeno, maleGeno, matingWeight)).
In this case, the mating weight only applies to the specific female X male mating
event. The two lists may NOT be mixed - the list must hold all length 2 or all
length 3 vectors.
}
