% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-SEMcisX_reduct1.R
\name{cubeSEMcisXreduct1}
\alias{cubeSEMcisXreduct1}
\title{Inheritance Cube: X-Linked ReMEDE (REpeat Mediated Excision of a Drive Element)}
\usage{
cubeSEMcisXreduct1(
  pF = 1,
  aF = 1,
  cF = 1,
  mmrF = 0,
  mmrM = 0,
  pDep = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)
}
\arguments{
\item{pF}{Rate of cleavage during GD process in females}

\item{aF}{Rate of cleavage during SEM process in females}

\item{cF}{Rate of "V" allele formation from SSA during SEM process in females}

\item{mmrF}{Rate of MMR in females, driving allelic conversion of "V" into "X"}

\item{mmrM}{Rate of MMR in males, driving allelic conversion of "V" into "X"}

\item{pDep}{Rate of cleavage during maternal deposition}

\item{eta}{Genotype-specific mating fitness}

\item{phi}{Genotype-specific sex ratio at emergence}

\item{omega}{Genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{Genotype-specific female pupatory success}

\item{xiM}{Genotype-specific male pupatory success}

\item{s}{Genotype-specific fractional reduction(increase) in fertility}
}
\value{
Named list containing the inheritance cube, transition matrix, genotypes,
wild-type allele, and all genotype-specific parameters.
}
\description{
The ReMEDE system, put forth by \href{GETLINKWHENEXISTS!!!}{Chennuri and Myles},
is a realized application of \href{https://doi.org/10.1098/rstb.2019.0804}{biodegradable gene drives}
with application of \href{https://doi.org/10.1016/j.celrep.2020.107841}{Small-Molecule Control}
for the removal of drive elements through induceable SSA.
}
\details{
This construct has a single, X-Linked target site that carries gRNA, Cas9, an
induceable endonuclease, and is flanked by direct repeats. This version has
been completely reduced, removing all resistance alleles. After that, there are
5 possible alleles:
\itemize{
 \item X: Wild-type X chromosome
 \item G: Active GD (gene drive), with inactive but functional SEM (self-elimination mechanism) element
 \item V: "wild-type", product of SEM, non-targetable by GD or SEM
 \item H: Active GD with active SEM
 \item Y: Y-chromosome, non-targetable
}

This provides a total genotype count of 14, however, there are only 10 viable
female genotypes and 4 viable male genotypes.

"V" alleles are simply a minor allele with the same protein sequence as the
major allele, "W". There is the possibility for allelic conversion of the "V"
allele into the "W" allele by mechanisms such as MMR.

This drive has female specific GD and SEM parameters, as well as \strong{correct}
maternal deposition. There are no dosage effects
modeled (i.e., having two GD alleles increasing or decreasing the GD rates).

Gene-drive and SEM parameters (p*, a*, c*, mmr*) are
all rates and values must fall in the range of [0, 1], inclusive.

Population parameters (phi, xiF, xiM) are either NULL, for default values, or
named vectors, where the names must match the names in the inheritance pattern.
These parameters are also rates and fall in the range of [0, 1], inclusive.

Omega and s may also be NULL or named vectors, but are weights, and thus fall
in the range [0, inf).

Eta, the male mating weights, may be NULL for default values, or applied as a
list of vectors. If the vectors are length 2 (e.g., c(maleGeno, matingWeight)),
then that weight is applied to all female genotypes for that specific male genotype.
Alternatively, vectors of length 3 may be supplied (e.g., c(femaleGeno, maleGeno, matingWeight)).
In this case, the mating weight only applies to the specific female X male mating
event. The two lists may NOT be mixed - the list must hold all length 2 or all
length 3 vectors.
}
