% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-SEM_reduct1.R
\name{cubeSEMreduct1}
\alias{cubeSEMreduct1}
\title{Inheritance Cube: ReMEDE (REpeat Mediated Excision of a Drive Element), Full Reduction}
\usage{
cubeSEMreduct1(
  pF = 1,
  aF = 1,
  cF = 1,
  pM = pF,
  aM = aF,
  cM = cF,
  mmrF = 0,
  mmrM = mmrF,
  pDep = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)
}
\arguments{
\item{pF}{Rate of cleavage during GD process in females}

\item{aF}{Rate of cleavage during SEM process in females}

\item{cF}{Rate of "V" allele formation from SSA during SEM process in females}

\item{pM}{Rate of cleavage during GD process in males}

\item{aM}{Rate of cleavage during SEM process in males}

\item{cM}{Rate of "V" allele formation from SSA during SEM process in males}

\item{mmrF}{Rate of MMR in females, driving allelic conversion of "V" into "W"}

\item{mmrM}{Rate of MMR in males, driving allelic conversion of "V" into "W"}

\item{pDep}{Rate of cleavage during maternal deposition}

\item{eta}{Genotype-specific mating fitness}

\item{phi}{Genotype-specific sex ratio at emergence}

\item{omega}{Genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{Genotype-specific female pupatory success}

\item{xiM}{Genotype-specific male pupatory success}

\item{s}{Genotype-specific fractional reduction(increase) in fertility}
}
\value{
Named list containing the inheritance cube, transition matrix, genotypes,
wild-type allele, and all genotype-specific parameters.
}
\description{
The ReMEDE system, put forth by \href{GETLINKWHENEXISTS!!!}{Chennuri and Myles},
is a realized application of \href{https://doi.org/10.1098/rstb.2019.0804}{biodegradable gene drives}
with application of \href{https://doi.org/10.1016/j.celrep.2020.107841}{Small-Molecule Control}
for the removal of drive elements through induceable SSA.
}
\details{
This construct has a single, autosomal target site that carries gRNA, Cas9,
an induceable endonuclease, and is flanked by direct repeats. This version
has been reduced, removing all resistance alleles. After that, there are 4 possible alleles:
\itemize{
 \item W: Wild-type
 \item G: Active GD (gene drive), with inactive but functional SEM (self-elimination mechanism) element
 \item V: "wild-type", product of SEM, non-targetable by GD or SEM
 \item H: Active GD with active SEM
}

"V" alleles are simply a minor allele with the same protein sequence as the
major allele, "W". There is the possibility for allelic conversion of the "V"
allele into the "W" allele by mechanisms such as MMR.

This drive has male and female specific GD and SEM parameters, but here are no dosage effects
modeled (i.e., having two GD alleles increasing or decreasing the GD rates).
Proper  maternal deposition is modeled, but with no resistance alleles possible.
}
