% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-SEMtrans.R
\name{cubeSEMtrans}
\alias{cubeSEMtrans}
\title{Inheritance Cube: ReMEDE (REpeat Mediated Excision of a Drive Element) in Trans}
\usage{
cubeSEMtrans(
  pF = 1,
  qF = 1,
  rF = 0,
  aF = 1,
  bF = 1,
  cF = 1,
  pM = pF,
  qM = qF,
  rM = rF,
  aM = aF,
  bM = bF,
  cM = cF,
  mmrF = 0,
  mmrM = mmrF,
  pDep = 0,
  qDep = 0,
  rDep = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)
}
\arguments{
\item{pF}{Rate of cleavage during GD process in females}

\item{qF}{Rate of HDR during GD process in females}

\item{rF}{Rate of in-frame resistance generation during GD process in females}

\item{aF}{Rate of cleavage during SEM process on GD allele in females}

\item{bF}{Rate of SSA during SEM process on GD allele in females}

\item{cF}{Rate of "V" allele formation from SSA during SEM process on GD allele in females}

\item{pM}{Rate of cleavage during GD process in males}

\item{qM}{Rate of HDR during GD process in males}

\item{rM}{Rate of in-frame resistance generation during GD process in males}

\item{aM}{Rate of cleavage during SEM process on GD allele in males}

\item{bM}{Rate of SSA during SEM process on GD allele in males}

\item{cM}{Rate of "V" allele formation from SSA during SEM process on GD allele in males}

\item{mmrF}{Rate of MMR in females, driving allelic conversion of "V" into "W"}

\item{mmrM}{Rate of MMR in males, driving allelic conversion of "V" into "W"}

\item{pDep}{Rate of cleavage during maternal deposition}

\item{qDep}{Rate of HDR during maternal deposition}

\item{rDep}{Rate of in-frame resistance generation during maternal deposition}

\item{eta}{Genotype-specific mating fitness}

\item{phi}{Genotype-specific sex ratio at emergence}

\item{omega}{Genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{Genotype-specific female pupatory success}

\item{xiM}{Genotype-specific male pupatory success}

\item{s}{Genotype-specific fractional reduction(increase) in fertility}

\item{xF}{Rate of cleavage during GD process on SEM allele in females}

\item{yF}{Rate of SSA during GD process on SEM allele in females}

\item{xM}{Rate of cleavage during GD process on SEM allele in males}

\item{yM}{Rate of SSA during GD process on SEM allele in males}
}
\value{
Named list containing the inheritance cube, transition matrix, genotypes,
wild-type allele, and all genotype-specific parameters.
}
\description{
This is the second ReMEDE system put forth by \href{GETLINKWHENEXISTS!!!}{Chennuri and Myles}.
It's another realization of \href{https://doi.org/10.1098/rstb.2019.0804}{biodegradable gene drives},
where the pieces exist as indeptendently-recombining constructs that only act
when in the presence of each other.
}
\details{
This construct consists of two autosomal target sites. The first site carries
gRNA and Cas9 for homing and removal of the SEM site. The second site consists
of an endonuclease targeting the Cas9 site. Both sites contain direct repeats
to induce SSA upon chromosome damage. There are 6 possible alleles at the first site:
\itemize{
 \item W: Wild-type allele
 \item G: Active GD (gene drive), with targetable SEM (self-elimination mechanism) site
 \item U: Low-cost resistant allele, non-targetable by GD or SEM (R1 in other literature)
 \item R: High-cost resistant allele, non-targetable by GD or SEM (R2 in other literature)
 \item V: "wild-type", product of SEM, non-targetable by GD or SEM
 \item S: Active GD with non-targetable SEM site
}

There are 4 possible alleles at the second site:
\itemize{
 \item W: Wild-type allele
 \item H: Active SEM element
 \item R: High-cost resistant allele, non-targetable by GD (R2 in other literature)
 \item E: "wild-type", produce of GD, non-targetbale by GD
}

This provides a total genotype count of 210. Genotypes are written (site 1)(site 1)(site 2)(site 2).
Therefore, a genotype of WWHR would be completely wild-type at the first locus,
and at the second locus a compound heterozygote with the SEM allele and a
costly resistance allele.

"V" and "E" alleles are simply minor alleles with the same protein sequence as the
major allele at their respective loci, "W". There is the possibility for allelic
conversion of the "V" allele into the "W" allele by mechanisms such as MMR.

This drive has male and female specific GD and SEM parameters, as well as
maternal deposition for both SEM and GD elements. There are no dosage effects
modeled (i.e., having two GD alleles increasing or decreasing the GD rates).
}
