% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-SEMtrans_reduct1.R
\name{cubeSEMtransReduct1}
\alias{cubeSEMtransReduct1}
\title{Inheritance Cube: ReMEDE (REpeat Mediated Excision of a Drive Element) in Trans}
\usage{
cubeSEMtransReduct1(
  pF = 1,
  aF = 1,
  cF = 1,
  xF = 1,
  pM = pF,
  aM = aF,
  cM = cF,
  xM = xF,
  mmrF = 0,
  mmrM = mmrF,
  pDep = 0,
  aDep = 0,
  cDep = 0,
  xDep = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)
}
\arguments{
\item{pF}{Rate of cleavage during GD process in females}

\item{aF}{Rate of cleavage during SEM process on GD allele in females}

\item{cF}{Rate of "V" allele formation from SSA during SEM process on GD allele in females}

\item{xF}{Rate of cleavage during GD process on SEM allele in females}

\item{pM}{Rate of cleavage during GD process in males}

\item{aM}{Rate of cleavage during SEM process on GD allele in males}

\item{cM}{Rate of "V" allele formation from SSA during SEM process on GD allele in males}

\item{xM}{Rate of cleavage during GD process on SEM allele in males}

\item{mmrF}{Rate of MMR in females, driving allelic conversion of "V" into "W"}

\item{mmrM}{Rate of MMR in males, driving allelic conversion of "V" into "W"}

\item{pDep}{Rate of cleavage during maternal deposition into W allele from GD}

\item{aDep}{Rate of cleavage during maternal deposition into G allele from SEM}

\item{cDep}{Rate of MMR, converting G into V, from maternal deposition from SEM}

\item{xDep}{Rate of cleavage of SEM allele from GD during maternal deposition}

\item{eta}{Genotype-specific mating fitness}

\item{phi}{Genotype-specific sex ratio at emergence}

\item{omega}{Genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{Genotype-specific female pupatory success}

\item{xiM}{Genotype-specific male pupatory success}

\item{s}{Genotype-specific fractional reduction(increase) in fertility}
}
\value{
Named list containing the inheritance cube, transition matrix, genotypes,
wild-type allele, and all genotype-specific parameters.
}
\description{
This is the second ReMEDE system put forth by \href{GETLINKWHENEXISTS!!!}{Chennuri and Myles}.
It's another realization of \href{https://doi.org/10.1098/rstb.2019.0804}{biodegradable gene drives},
where the pieces exist as independently-recombining constructs that only act
when in the presence of each other.
}
\details{
This construct consists of two autosomal target sites. The first site carries
gRNA and Cas9 for homing and removal of the SEM site. The second site consists
of an endonuclease targeting the Cas9 site. Both sites contain direct repeats
to induce SSA upon chromosome damage. This version has
been slightly reduced, removing all resistance alleles.
After that, there are 3 possible alleles at the first site:
\itemize{
 \item W: Wild-type allele
 \item G: Active GD (gene drive), with targetable SEM (self-elimination mechanism) site
 \item V: "wild-type", product of SEM, non-targetable by GD or SEM
}

There are 3 possible alleles at the second site:
\itemize{
 \item W: Wild-type allele
 \item H: Active SEM element
 \item E: "wild-type", produce of GD, non-targetbale by GD
}

This provides a total genotype count of 36. Genotypes are written (site 1)(site 1)(site 2)(site 2).
Therefore, a genotype of WWHE would be completely wild-type at the first locus,
and at the second locus a compound heterozygote with the SEM allele and a
non-targetable wild-type equivalent allele.

"V" and "E" alleles are simply minor alleles with the same protein sequence as the
major allele at their respective loci, "W". There is the possibility for allelic
conversion of the "V" allele into the "W" allele by mechanisms such as MMR.

This drive has male and female specific GD and SEM parameters, as well as
maternal deposition for both SEM and GD elements. There are no dosage effects
modeled (i.e., having two GD alleles increasing or decreasing the GD rates).

Gene-drive and SEM parameters (p*, a*, c*, x*, mmr*) are
all rates and values must fall in the range of [0, 1], inclusive.

Population parameters (phi, xiF, xiM) are either NULL, for default values, or
named vectors, where the names must match the names in the inheritance pattern.
These parameters are also rates and fall in the range of [0, 1], inclusive.

Omega and s may also be NULL or named vectors, but are weights, and thus fall
in the range [0, inf).

Eta, the male mating weights, may be NULL for default values, or applied as a
list of vectors. If the vectors are length 2 (e.g., c(maleGeno, matingWeight)),
then that weight is applied to all female genotypes for that specific male genotype.
Alternatively, vectors of length 3 may be supplied (e.g., c(femaleGeno, maleGeno, matingWeight)).
In this case, the mating weight only applies to the specific female X male mating
event. The two lists may NOT be mixed - the list must hold all length 2 or all
length 3 vectors.
}
