% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-epiImperial.R
\name{equilibrium_Imperial_decoupled}
\alias{equilibrium_Imperial_decoupled}
\title{Calculate Equilibrium for Imperial Model}
\usage{
equilibrium_Imperial_decoupled(age_vector, ft, eir, theta, cube, spn_P)
}
\arguments{
\item{age_vector}{age structure of population (see vignette for example)}

\item{ft}{proportion of population seeking treatment}

\item{eir}{desired annual EIR}

\item{theta}{parameters}

\item{cube}{inheritance cube}

\item{spn_P}{places of the stochastic petri net}
}
\value{
a matrix of the equilibrium number of humans in each Imperial stage by
age, and immunity. mosquito equilibrium values, and full theta vector
}
\description{
This function calculates the human and mosquito equilibrium values for the
decoupled Imperial model. \strong{Currently this only works in one node.}
}
\details{
For mosquito equilibrium, this function calls \code{\link{equilibrium_SEI_Imperial_mosy}} internally.

For human equilibrium, this function calls \code{\link{equilibrium_Imperial_decoupled_human}} internally.

For examples of using this function, see:
\code{vignette("epi-node-imperial", package = "MGDrivE2")}
}
