% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-epi-decoupled.R
\name{equilibrium_Imperial_decoupled_human}
\alias{equilibrium_Imperial_decoupled_human}
\title{This function calculates the human equilibrium values for the decoupled Imperial model.
Requires the age structure of the population
Currently this only works in one node.}
\usage{
equilibrium_Imperial_decoupled_human(age_vector, ft, EIR, model_param_list)
}
\arguments{
\item{age_vector}{age structure of population (see vignette for example)}

\item{ft}{proportion of population seeking treatment}

\item{EIR}{desired annual EIR}

\item{model_param_list}{parameters for the}
}
\value{
a matrix of the equilibrium number of humans in each Imperial stage by age, and immunity
}
\description{
This function calculates the human equilibrium values for the decoupled Imperial model.
Requires the age structure of the population
Currently this only works in one node.
}
