% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-epi-decoupled.R
\name{equilibrium_SEI_decoupled}
\alias{equilibrium_SEI_decoupled}
\title{Calculate Decoupled Equilibrium}
\usage{
equilibrium_SEI_decoupled(params, spn_P, cube)
}
\arguments{
\item{params}{a named list of parameters}

\item{spn_P}{the set of places (P)}

\item{cube}{an inheritance cube from the \code{MGDrivE} package (e.g. \code{\link[MGDrivE]{cubeMendelian}})}
}
\value{
a list of parameters and initial conditions
}
\description{
This function calculates the mosquito and human equilibrium for the decoupled
simulations. \strong{This function only runs 1 node}. It cannot generate
anything for a network.
}
\details{
For mosquito equilibrium, this function calls \code{\link{equilibrium_SEI_decoupled_mosy}} internally.

For human equilibrium, this function calls \code{\link{equilibrium_SEI_decoupled_human}} internally.

For examples of using this function, see:
\code{vignette("epi-node-decoupled", package = "MGDrivE2")}
}
