% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-epi-decoupled.R
\name{equilibrium_SEI_decoupled_human}
\alias{equilibrium_SEI_decoupled_human}
\title{This function calculates the equilibrium values for the decoupled SIS human states.
Currently this only works in one node.}
\usage{
equilibrium_SEI_decoupled_human(params)
}
\arguments{
\item{params}{a named list of parameters (see details)}
}
\value{
a vector of the equilibrium number of humans in each SIS stage
}
\description{
This function calculates the equilibrium values for the decoupled SIS human states.
Currently this only works in one node.
}
