% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium-epiSEIR.R
\name{equilibrium_SEI_SEIR}
\alias{equilibrium_SEI_SEIR}
\title{Calculate Equilibrium for Mosquito SEI - Human SEIR Model}
\usage{
equilibrium_SEI_SEIR(
  params,
  node_list = "b",
  NF = NULL,
  phi = 0.5,
  NH = NULL,
  log_dd = TRUE,
  spn_P,
  pop_ratio_Aq = NULL,
  pop_ratio_F = NULL,
  pop_ratio_M = NULL,
  pop_ratio_H = c(1, 0, 0, 0),
  cube
)
}
\arguments{
\item{params}{a named list of parameters (see details)}

\item{node_list}{a character vector specifying what type of nodes to create;
(m = a node with only mosquitoes, h = a node with only humans, b = a node with both humans and mosquitoes)}

\item{NF}{vector of female mosquitoes at equilibrium, for mosquito-only nodes}

\item{phi}{sex ratio of mosquitoes at emergence}

\item{NH}{vector of humans at equilibrium, for human-only nodes}

\item{log_dd}{Boolean: TRUE implies logistic density dependence, FALSE implies Lotka-Volterra model}

\item{spn_P}{the set of places (P) (see details)}

\item{pop_ratio_Aq}{May be empty; if not, a named vector or matrix. (see details)}

\item{pop_ratio_F}{May be empty; if not, a named vector or matrix. (see details)}

\item{pop_ratio_M}{May be empty; if not, a named vector or matrix. (see details)}

\item{pop_ratio_H}{Prevalence in human-only nodes, default is all susceptible}

\item{cube}{an inheritance cube from the \code{MGDrivE} package (e.g. \code{\link[MGDrivE]{cubeMendelian}})}
}
\value{
a vector of the equilibrium number of females in each SEI stage
}
\description{
Given prevalence of disease in humans (modeled as an SEIR: Susceptible-Latent-Infected-Recovered
process with birth and death) and entomological parameters of transmission, this
function calculates the quasi-stationary distribution of adult female mosquitoes
across SEI (Susceptible-Exposed-Infectious) stages, allowing for Erlang distributed E stage.
}
\details{
This function handles 3 types of nodes: Human only, mosquito only, and nodes
with both. These nodes are set using the \code{node_list} parameter.
Mosquito-only node equilibrium calls \code{\link{equilibrium_lifeycle}}, which
follows one of two models: classic logistic dynamics or the Lotka-Volterra
competition model. This is determined by the parameter \code{log_dd}, and it
changes elements of the return list: \code{K} is returned for logistic dynamics,
or \code{gamma} is returned for Lotka-Volterra dynamics. This
is parameterized with the \code{NF} parameter to define the adult female numbers.
This parameter only needs to be supplied if there are mosquito-only nodes.

Human-only nodes don't require any equilibrium calculations. These nodes use
the \code{NH} and \code{pop_ratio_H} to set adult human populations and
infection rates in nodes. These two parameters only need to be supplied
if there are human-only nodes. \code{pop_ratio_H} needs to be a matrix with the
number of rows equal to the number of human-only patches, and 4 columns. The
columns are assumed to be fractions of the population in "S", "E", "I", or "R"
states, and every row must sum to 1.

For human and mosquito nodes, this function calls \code{\link{make_Q_SEI}} to construct the
infinitesimal generator matrix which is used to solve for the quasi-stationary
(stochastic) or equilibrium (deterministic) distribution of mosquitoes over stages.
Parameters are provided by \code{params}.

For information on the method used to solve this distribution, see section
"3.1.3 Nonsingularity of the Subintensity Matrix" of:
\itemize{
\item Bladt, Mogens, and Bo Friis Nielsen. Matrix-exponential distributions in
applied probability. Vol. 81. New York: Springer, 2017.
}

The places (\code{spn_P}) object is generated from one of the following:
\code{\link{spn_P_lifecycle_node}}, \code{\link{spn_P_lifecycle_network}},
\code{\link{spn_P_epiSIS_node}}, \code{\link{spn_P_epiSIS_network}},
\code{\link{spn_P_epiSEIR_node}}, or \code{\link{spn_P_epiSEIR_network}}.

The initial population genotype ratios are set by supplying the \code{pop_ratio_Aq},
\code{pop_ratio_F}, and \code{pop_ratio_M} values. The default value is NULL,
and the function will use the wild-type alleles provided in the \code{cube}
object. However, one can supply
several different objects to set the initial genotype ratios. All genotypes provided
must exist in the \code{cube} (this is checked by the function). If a single, named vector
is provided, then all patches will be initialized with the same ratios. If a
matrix is provided, with the number of columns (and column names) giving the
initial genotypes, and a row for each patch, each patch can be set to a different
initial ratio. The three parameters do not need to match each other.

The \code{params} argument supplies all of the ecological and epidemiological
parameters necessary to calculate equilibrium values. This is used to set the
initial population distribution and during the simulation to maintain equilibrium.
This \code{params} must include the following named parameters, noted as being
the same as lifecycle parameters, or new for the epidemiological equilibrium
\itemize{
\item \strong{(Lifecycle parameters)}
\itemize{
\item \code{qE}: inverse of mean duration of egg stage
\item \code{nE}: shape parameter of Erlang-distributed egg stage
\item \code{qL}: inverse of mean duration of larval stage
\item \code{nL}: shape parameter of Erlang-distributed larval stage
\item \code{qP}: inverse of mean duration of pupal stage
\item \code{nP}: shape parameter of Erlang-distributed pupal stage
\item \code{muE}: egg mortality
\item \code{muL}: density-independent larvae mortality
\item \code{muP}: pupae mortality
\item \code{muF}: adult female mortality
\item \code{muM}: adult male mortality
\item \code{beta}: egg-laying rate, daily
\item \code{nu}: mating rate of unmated females
}
\item \strong{(Epidemiological parameters)}
\itemize{
\item \code{NH}: number of humans, can be a vector
\item \code{X}: SEIR prevalence in humans, can be a vector of length 4 for 1 node, or a matrix for many nodes
\item \code{NFX}: number of female mosquitoes, only required if any prevalence (X) is zero
\item \code{b}: mosquito to human transmission efficiency, can be a vector
\item \code{c}: human to mosquito transmission efficiency, can be a vector
\item \code{r}: rate of recovery in humans (1/duration of infectiousness)
\item \code{muH}: death rate of humans (1/avg lifespan)
\item \code{f}: rate of blood feeding
\item \code{Q}: human blood index
\item \code{qEIP}: related to scale parameter of Gamma distributed EIP (1/qEIP is mean length of EIP)
\item \code{nEIP}: shape parameter of Gamma distributed EIP
\item \code{delta}: inverse duration of the latent stage (E)
}
}

The return list contains all of the parameters necessary later in the simulations.

For equilibrium without epidemiological parameters, see \code{\link{equilibrium_lifeycle}}.
For equilibrium without latent humans (SIS dynamics), see \code{\link{equilibrium_SEI_SIS}}.
}
