% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-auxiliary.R
\name{get_shape}
\alias{get_shape}
\title{Calculate Erlang shape parameter}
\usage{
get_shape(cv, q)
}
\arguments{
\item{cv}{coefficient of variation (CV) between mean and standard deviation of dwell times,
smaller values of CV correspond to distributions less dispersed around their mean and larger
value to more dispersed distributions.}

\item{q}{inverse of mean dwell time}
}
\value{
integer value representing the coefficient of variation in Erlang-distributed life stages.
}
\description{
Calculate Erlang shape parameter
}
