% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imperial-ode.R
\name{human_Imperial_ODE}
\alias{human_Imperial_ODE}
\title{ODE describing the age-structured Imperial model
used in decoupled sampling, which will pass in values of I_V and
return the human states for usein the mosquito portion of the model}
\usage{
human_Imperial_ODE(t, state, parameters)
}
\arguments{
\item{t}{starting time of simulation}

\item{state}{distributon of disease states}

\item{parameters}{parameter set}
}
\value{
matrix of disease states after integration
}
\description{
ODE describing the age-structured Imperial model
used in decoupled sampling, which will pass in values of I_V and
return the human states for usein the mosquito portion of the model
}
