% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-ode-decoupled.R
\name{human_SIS_ODE}
\alias{human_SIS_ODE}
\title{ODE describing the basic SIS model
used in decoupled sampling, which will pass in values of I_V and
return the human states for use in the mosquito portion of the model}
\usage{
human_SIS_ODE(t, state, parameters)
}
\arguments{
\item{t}{starting time of simulation}

\item{state}{distribution of disease states}

\item{parameters}{parameter set}
}
\value{
list of disease states after integration
}
\description{
ODE describing the basic SIS model
used in decoupled sampling, which will pass in values of I_V and
return the human states for use in the mosquito portion of the model
}
