% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu_ts}
\alias{mu_ts}
\title{Mosquito Death Rates, Comoros Islands}
\format{
matrix with 3 named columns and 8760 rows:
\describe{
\item{Grande_Comore}{Hourly death rates for main island}
\item{Moheli}{Hourly death rates for second island}
\item{Anjouan}{Hourly death rates for smallest island}
}
}
\usage{
data(mu_ts)
}
\description{
This is a matrix containing estimated mosquito death rates from the Comoros
Islands, between Mozambique and Madagascar. It provides hourly death rates over
the course of one year.
}
\keyword{datasets}
