% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-trajectory-decoupled-CSV.R
\name{sim_trajectory_CSV_decoupled}
\alias{sim_trajectory_CSV_decoupled}
\title{Simulate Trajectory From a SPN Model}
\usage{
sim_trajectory_CSV_decoupled(
  x0,
  h0,
  inf_labels,
  tmax,
  dt = 1,
  dt_stoch = 0.1,
  folders = "./",
  S,
  hazards,
  SPN_P,
  theta,
  Sout = NULL,
  sampler = "tau",
  method = "lsoda",
  events = NULL,
  batch = NULL,
  verbose = TRUE,
  human_ode = "Imperial",
  cube = cube,
  ...
)
}
\arguments{
\item{x0}{the initial marking of the SPN (initial state, M0)}

\item{h0}{the initial human state distribution}

\item{inf_labels}{labels corresponding to female mosquito infection hazard}

\item{tmax}{the final time to end simulation}

\item{dt}{the time-step at which to return output (\strong{not} the time-step of the sampling algorithm)}

\item{dt_stoch}{time-step used for approximation of hazards}

\item{folders}{vector of folders to write output}

\item{S}{a stoichiometry \code{\link[Matrix]{Matrix-class}} object}

\item{hazards}{list of hazard functions}

\item{SPN_P}{stochastic petri net places}

\item{theta}{parameters}

\item{Sout}{an optional matrix to track event firings}

\item{sampler}{determines sampling algorithm, one of; "ode", "tau", "cle", or "dm"}

\item{method}{if \code{sampler} is "ode", the solver to use, from \code{deSolve}}

\item{events}{a \code{data.frame} of events}

\item{batch}{a \code{list} of batch migration events, created from \code{\link[MGDrivE2]{batch_migration}}, may be set to \code{NULL} if not used}

\item{verbose}{print a progress bar?}

\item{human_ode}{human ode function}

\item{cube}{inheritance cube}

\item{...}{further named arguments passed to the step function}
}
\value{
NULL - prints output to .csv files
}
\description{
This function provides a unified interface to the various simulation algorithms
for SPN, returning output sampled at a lattice of time points to the user, and
handling various exogenous events that may occur during the simulation
(such as release of adult mosquitoes). This function is used in decoupled sampling, where
the mosquito and human states are separated. This is used primarily when using the
Imperial model of malaria transmission.
}
\details{
\code{dt_stoch} is used by the Poisson Time-Step (\code{\link{step_PTS}}) and
Chemical Langevin (\code{\link{step_CLE}}) methods to approximate the hazards.
A smaller \code{dt_stoch} provides a better approximation, but will take longer
to run.

The stoichiometry matrix (\code{S}) is generated in \code{\link{spn_S}}.

The list of hazards (\code{hazards}) come from \code{\link{spn_hazards}}.

Several samplers are provided. The default is a Poisson Time-Step
(\code{\link{step_PTS}}) method. Other options are Gillespie's Direct Method
(\code{\link{step_DM}}) and a Chemical Langevin sampler (\code{\link{step_CLE}}).
Additionally, for convenience, an ODE "sampler" (\code{\link{step_ODE}}) is
provided for compatibility with other samplers. This function uses methods from
\code{deSolve}.

If using the \code{ode} sampler, several \code{methods} are provided in the \code{deSolve}
package, see \code{\link[deSolve]{ode}}. For inhomogeneous systems, consider
using the "rk4" method to avoid excessive integration times.

Additionally, \code{events} objects must follow the format required by
\code{deSolve}. This was done for consistency, see \code{\link[deSolve]{events}}
for more information.

This function writes all output to .csv files. Each simulation is written to
a \code{folder} element - the number of repetitions is the number of folders
provided. For now, only adult mosquito states, human states, clinical incidence, and pathogen prevalence are
written to CSVs.

This function tracks state variables specified by argument \code{stage} by default; an optional argument \code{Sout}
can be provided to track number of event firings each time step (for discrete stochastic simulations),
or cumulative intensity (for continuous stochastic simulations), or the rate function of
particular events for ODE simulation. The matrix must have number of columns equal to
number of events in the system (the number of hazard functions), and a row for each tracking
variable. If \code{Sout} is provided, it outputs an additional csv, "Tracking.csv".
The function \code{\link{track_hinf}} is provided, which builds a matrix to track
human infection events.

To return simulations to R for further processing, see \code{\link{sim_trajectory_R}}.
}
