% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PN-epiSEIR-node-T.R
\name{spn_T_epiSEIR_node}
\alias{spn_T_epiSEIR_node}
\title{Make Transitions (T) For a Node (SEI Mosquitoes - SEIR Humans)}
\usage{
spn_T_epiSEIR_node(spn_P, params, cube)
}
\arguments{
\item{spn_P}{set of places produced by \code{\link{spn_P_epiSEIR_node}}}

\item{params}{a named list of parameters (see details)}

\item{cube}{an inheritance cube from the \code{MGDrivE} package (e.g. \code{\link[MGDrivE]{cubeMendelian}})}
}
\value{
a list with two elements: \code{T} contains transitions packets as lists,
\code{v} is the character vector of transitions (T)
}
\description{
This function makes the set of transitions (T) for a SPN. It is used alone
if our model is a single-node metapopulation of mosquito and human dynamics;
otherwise it is used as part of other functions to make SPN models with larger
state spaces (metapopulation models, see \code{\link{spn_T_epiSEIR_network}}).
}
\details{
This function takes the places produced from \code{\link{spn_P_epiSEIR_node}}
and builds all possible transitions between subsets of those places.

The \code{params} argument supplies all of the ecological parameters necessary
to calculate equilibrium values. This function requires the \code{nE},
\code{nL}, \code{nP}, and \code{nEIP} parameters to be specified. For more details, see
\code{\link{equilibrium_SEI_SEIR}}

While this function produces all structural information related to transitions,
hazards are produced by a separate function, \code{\link{spn_hazards}}.
}
\examples{
\dontrun{
  # parameters, see vignette MGDrivE2 SEIR Epidemiological Dynamics
  theta <- list(qE = 1/4, nE = 2, qL = 1/3, nL = 3, qP = 1/6, nP = 2,
                muE = 0.05, muL = 0.15, muP = 0.05, muF = 0.09, muM = 0.09,
                beta = 16, nu = 1/(4/24), NH = 250, X = c(1,0,0,0), NFX = 500,
                f = 1/3, Q = 0.9, b = 0.55, c = 0.15, delta = 1/5, r = 1/14,
                muH = 1/(62*365), qEIP = 1/11, nEIP = 3 )

  # spn_P needs setup elsewhere, see vignettes

  spn_T <- spn_T_epiSEIR_node(spn_P = spn_P, params = theta,
                              cube = MGDrivE::cubeMendelian() )
}

}
