% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-plot.R
\name{summarize_females_epi}
\alias{summarize_females_epi}
\title{Summarize Adult Females (One Node or Metapopulation Network, SEI Mosquitoes)}
\usage{
summarize_females_epi(out, spn_P)
}
\arguments{
\item{out}{the output of \code{\link[MGDrivE2]{sim_trajectory_R}}}

\item{spn_P}{the places of the SPN, see details}
}
\value{
a 4 to 6 column dataframe for plotting with ggplot2
}
\description{
For \code{MGDrivE2} simulations of mosquito epidemiological dynamics in a single
node or metapopulation network, this function sums over the male mate genotype
as well as EIP bins to get population trajectories of adult female mosquitoes
by their genotype and (S,E,I) status.
}
\details{
The places (\code{spn_P}) object is generated from one of the following:
\code{\link{spn_P_epiSIS_node}}, \code{\link{spn_P_epiSIS_network}},
\code{\link{spn_P_epiSEIR_node}}, or \code{\link{spn_P_epiSEIR_network}}.

The return object depends on the data provided. If the simulation was only 1 node,
then no \code{node} designation is returned. If only one repetition was performed,
no \code{rep} designation is returned. Columns always returned include: \code{time},
\code{inf}, \code{genotype}, and \code{value}.

For examples of using this function, this or any vignette which simulates epi dynamics:
\code{vignette("epi-node", package = "MGDrivE2")}
}
