% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-plot.R
\name{summarize_humans_epiImperial}
\alias{summarize_humans_epiImperial}
\title{Summarize Humans for Imperial Model}
\usage{
summarize_humans_epiImperial(out, index = 1)
}
\arguments{
\item{out}{the output of \code{\link[MGDrivE2]{sim_trajectory_R}}}

\item{index}{the desired age compartment for which to pull trajectory}
}
\value{
dataframe for plotting with ggplot2
}
\description{
The Imperial model outputs six human states for each age compartment.
This function accepts the output matrix and the desired index of an age compartment
and returns the trajectory of all human states in that given age compartment (default 1)
}
