% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-plot.R
\name{summarize_males}
\alias{summarize_males}
\title{Summarize Adult Males (One Node or Metapopulation Network)}
\usage{
summarize_males(out)
}
\arguments{
\item{out}{the output of \code{\link[MGDrivE2]{sim_trajectory_R}}}
}
\value{
a 3 to 5 column dataframe for plotting with ggplot2
}
\description{
For \code{MGDrivE2} simulations of mosquito lifecycle dynamics or human infection
dynamics, in a node or metapopulation network, this function summarizes
population trajectories of adult male mosquitoes by their genotype.
}
\details{
The return object depends on the data provided. If the simulation was only 1 node,
then no \code{node} designation is returned. If only one repetition was performed,
no \code{rep} designation is returned. Columns always returned include: \code{time},
\code{genotype}, and \code{value}.

For examples of using this function, this or any vignette which visualizes output:
\code{vignette("lifecycle-node", package = "MGDrivE2")}
}
