% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-tracking.R
\name{track_hinf}
\alias{track_hinf}
\title{Make tracking matrix for human infection events}
\usage{
track_hinf(spn_T, S)
}
\arguments{
\item{spn_T}{set of transitions}

\item{S}{stoichiometry matrix}
}
\value{
a \code{\link[Matrix]{sparseMatrix}} object
}
\description{
Create a matrix object for tracking incidence in human population
to be passed to either \code{\link{sim_trajectory_CSV}} or \code{\link{sim_trajectory_R}}.
}
\details{
The returned matrix can be passed to the \code{Sout} argument of \code{\link{sim_trajectory_CSV}} or \code{\link{sim_trajectory_R}}.
}
