% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PN-lifecycle-network-T.R
\name{spn_T_lifecycle_network}
\alias{spn_T_lifecycle_network}
\title{Make Transitions (T) For a Network (Mosquitoes only)}
\usage{
spn_T_lifecycle_network(spn_P, params, cube, m_move)
}
\arguments{
\item{spn_P}{set of places produced by \code{\link{spn_P_lifecycle_network}}}

\item{params}{a named list of parameters (see details)}

\item{cube}{an inheritance cube from the \code{MGDrivE} package (e.g. \code{\link[MGDrivE]{cubeMendelian}})}

\item{m_move}{binary adjacency matrix indicating if movement of mosquitoes between nodes is possible or not}
}
\value{
a list with two elements: \code{T} contains transitions packets as lists,
\code{v} is the character vector of transitions (T)
}
\description{
This function makes the set of transitions (T) for a SPN model of a
metapopulation network. It is the network version of \code{\link{spn_T_lifecycle_node}}.
}
\details{
This function takes the places produced from \code{\link{spn_P_lifecycle_network}}
and builds all possible transitions between subsets of those places.

The \code{params} argument supplies all of the ecological parameters necessary
to calculate equilibrium values. This function requires the \code{nE},
\code{nL}, and \code{nP} parameters to be specified. For more details, see
\code{\link{equilibrium_lifeycle}}

While this function produces all structural information related to transitions,
hazards are produced by a separate function, \code{\link{spn_hazards}}.

For larger networks, this function may take some time to return, please be patient;
the Petri Net modeling formalism trades additional computation time at model
initialization for faster sampling of trajectories within a simulation.

Please note, the movement matrix (\code{m_move}) is NOT a
stochastic matrices, just a binary matrix that say if i,j can exchange population.
Diagonal elements must be \code{FALSE}.
}
\examples{
\dontrun{
  # parameters, see vignette MGDrivE2: One Node Lifecycle Dynamics
  theta <- list(qE = 1/4, nE = 2, qL = 1/3, nL = 3, qP = 1/6, nP = 2,
                muE = 0.05, muL = 0.15, muP = 0.05, muF = 0.09, muM = 0.09,
                beta = 16, nu = 1/(4/24) )

  # spn_P needs setup elsewhere, see vignettes

  # setup 2-node allowed migrations
  #  let node 1 migrate to node 2, but not vice-versa
  m_move <- matrix(data = FALSE, nrow = 2, ncol = 2)
  m_move[1,2] <- TRUE

  spn_T <- spn_T_lifecycle_network(spn_P = spn_P, params = theta,
                                   cube = MGDrivE::cubeMendelian(),
                                   m_move = m_move )
}

}
